try more nuanced fix to fsevent deployment issue on non macOs platforms
It looks like the pattern is to specify fsevents as an optionalDependencies
in the package.json.  If it's incompatible with your platform, it will
be ignored like this:

Running heroku-postbuild (yarn)
...
[3/5] Fetching packages...
remote: info fsevents@2.1.3: The platform "linux" is incompatible with this module.
remote: info "fsevents@2.1.3" is an optional dependency and failed compatibility check. Excluding it from installation.

So I'm hoping that the combo of optionalDependencies and resolutions
play nicely together on my hosting target.